requirements.txt
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
openai>=1.0.0
urllib3>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
google-cloud-storage>=2.10.0

